+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 # Changelog
 
 # Todo
-- sqlmath - Update sql-function win_cosfit2_step() to be able to update cosine-fit from last datapoint.
 - none
 
 # v2023.8.1-beta
+- sqlmath - Update sql-function win_cosfit2_step() to be able to update cosine-fit as well.
 - sqlmath - merge c-struct WinCosfitInternal, WinCosfitResult into c-struct WinCosfit. - part1
 - sqlmath - Update sql-functions win_cosfit2(), win_emax(), win_quantilex() to move variable-length arguments to last position.
 - sqlmath - Update sql-function win_cosfit2() with additional argument modeNoCsf.